refactor: clean up updateMoveToml function (#147)
