+++ .changeset/eleven-hotels-tease.md
@@ -1,0 +1,5 @@
+---
+'@axelar-network/axelar-cgp-sui': patch
+---
+
+refactor: simplified updateMoveToml
