+++ package.json
@@ -43,6 +43,7 @@
     "@mysten/sui": "^1.3.0",
     "ethers": "^5.0.0",
     "secp256k1": "^5.0.0",
+    "smol-toml": "^1.3.0",
     "typescript": "^5.3.3"
   },
   "devDependencies": {
