+++ src/utils.ts
@@ -1,80 +1,31 @@
 import { execSync } from 'child_process';
 import fs from 'fs';
 import { getFullnodeUrl } from '@mysten/sui/client';
-
-export function getModuleNameFromSymbol(symbol: string) {
-    function isNumber(char: string) {
-        return char >= '0' && char <= '9';
-    }
-
-    function isLowercase(char: string) {
-        return char >= 'a' && char <= 'z';
-    }
-
-    function isUppercase(char: string) {
-        return char >= 'A' && char <= 'Z';
-    }
-
-    let i = 0;
-    const length = symbol.length;
-    let moduleName = '';
-
-    while (isNumber(symbol[i])) {
-        i++;
-    }
-
-    while (i < length) {
-        const char = symbol[i];
-
-        if (isLowercase(char) || isNumber(char)) {
-            moduleName += char;
-        } else if (isUppercase(char)) {
-            moduleName += char.toLowerCase();
-        } else if (char === '_' || char === ' ') {
-            moduleName += '_';
-        }
-
-        i++;
-    }
-
-    return moduleName;
-}
+import toml from 'smol-toml';
 
 export function updateMoveToml(packageName: string, packageId: string, moveDir: string = `${__dirname}/../move`) {
-    const path = `${moveDir}/${packageName}/Move.toml`;
+    // Path to the Move.toml file for the package
+    const movePath = `${moveDir}/${packageName}/Move.toml`;
 
-    let toml = fs.readFileSync(path, 'utf8');
-
-    const lines = toml.split('\n');
-
-    const versionLineIndex = lines.findIndex((line: string) => line.slice(0, 7) === 'version');
-
-    if (!(lines[versionLineIndex + 1].slice(0, 12) === 'published-at')) {
-        lines.splice(versionLineIndex + 1, 0, '');
+    // Check if the Move.toml file exists
+    if (!fs.existsSync(movePath)) {
+        throw new Error(`Move.toml file not found for given path: ${movePath}`);
     }
 
-    lines[versionLineIndex + 1] = `published-at = "${packageId}"`;
+    // Read the Move.toml file
+    const moveRaw = fs.readFileSync(movePath, 'utf8');
 
-    const addressesIndex = lines.findIndex((line: string) => line.slice(0, 11) === '[addresses]');
+    // Parse the Move.toml file as JSON
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const moveJson = toml.parse(moveRaw) as any;
 
-    for (let i = addressesIndex + 1; i < lines.length; i++) {
-        const line = lines[i];
-        const eqIndex = line.indexOf('=');
+    // Update the published-at field under the package section e.g. published-at = "0x01"
+    moveJson.package['published-at'] = packageId;
 
-        if (
-            eqIndex < 0 ||
-            line.slice(0, packageName.length) !== packageName ||
-            line.slice(packageName.length, eqIndex) !== Array(eqIndex - packageName.length + 1).join(' ')
-        ) {
-            continue;
-        }
+    // Update the package address under the addresses section e.g. gas_service = "0x1"
+    moveJson.addresses[packageName] = packageId;
 
-        lines[i] = line.slice(0, eqIndex + 1) + ` "${packageId}"`;
-    }
-
-    toml = lines.join('\n');
-
-    fs.writeFileSync(path, toml);
+    fs.writeFileSync(movePath, toml.stringify(moveJson));
 }
 
 export function copyMovePackage(packageName: string, fromDir: null | string, toDir: string) {
